+++ cmd/status.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/Clever/microplane/clone"
 	"github.com/Clever/microplane/initialize"
+	"github.com/Clever/microplane/lib"
 	"github.com/Clever/microplane/merge"
 	"github.com/Clever/microplane/plan"
 	"github.com/Clever/microplane/push"
@@ -33,30 +34,22 @@ var statusCmd = &cobra.Command{
 			log.Fatalf("error loading init.json: %s\n", err.Error())
 		}
 
-		singleRepo, err := cmd.Flags().GetString("repo")
-		if err == nil && singleRepo != "" {
-			valid := false
-			validRepoNames := []string{}
-			for _, r := range initOutput.Repos {
-				if r.Name == singleRepo {
-					valid = true
-					break
-				}
-				validRepoNames = append(validRepoNames, r.Name)
+		repos, err := whichRepos(cmd)
+		if err != nil {
+			log.Fatal(err)
+		}
+		sync, err := cmd.Flags().GetBool("sync")
+		if err != nil {
+			log.Fatal(err)
+		}
+		if sync {
+			err = parallelize(repos, syncOneRepo)
+			if err != nil {
+				// TODO: dig into errors and display them with more detail
+				log.Fatal(err)
 			}
-			if !valid {
-				log.Fatalf("%s not a targeted repo name (valid target repos are: %s)", singleRepo, strings.Join(validRepoNames, ", "))
-			}
-			isSingleRepo = true
 		}
 
-		repos := []string{}
-		for _, r := range initOutput.Repos {
-			if singleRepo != "" && r.Name != singleRepo {
-				continue
-			}
-			repos = append(repos, r.Name)
-		}
 		printStatus(repos)
 	},
 }
@@ -76,7 +69,7 @@ func joinWithTab(s ...string) string {
 	return strings.Join(s, "\t")
 }
 
-func printStatus(repos []string) {
+func printStatus(repos []lib.Repo) {
 	out := tabWriterWithDefaults()
 	fmt.Fprintln(out, joinWithTab("REPO", "STATUS", "DETAILS"))
 	for _, r := range repos {
@@ -86,19 +79,20 @@ func printStatus(repos []string) {
 		if len(d3) > 150 {
 			d3 = d3[:150] + "..."
 		}
-		fmt.Fprintln(out, joinWithTab(r, status, d3))
+		fmt.Fprintln(out, joinWithTab(r.Name, status, d3))
 	}
 	out.Flush()
 }
 
-func getRepoStatus(repo string) (status, details string) {
+func getRepoStatus(repo lib.Repo) (status, details string) {
+	repoName := repo.Name
 	status = "initialized"
 	details = ""
 	var cloneOutput struct {
 		clone.Output
 		Error string
 	}
-	if !(loadJSON(outputPath(repo, "clone"), &cloneOutput) == nil && cloneOutput.Success) {
+	if !(loadJSON(outputPath(repoName, "clone"), &cloneOutput) == nil && cloneOutput.Success) {
 		if cloneOutput.Error != "" {
 			details = color.RedString("(clone error) ") + cloneOutput.Error
 		}
@@ -110,7 +104,7 @@ func getRepoStatus(repo string) (status, details string) {
 		plan.Output
 		Error string
 	}
-	if !(loadJSON(outputPath(repo, "plan"), &planOutput) == nil && planOutput.Success) {
+	if !(loadJSON(outputPath(repoName, "plan"), &planOutput) == nil && planOutput.Success) {
 		if planOutput.Error != "" {
 			details = color.RedString("(plan error) ") + planOutput.Error
 		}
@@ -129,7 +123,7 @@ func getRepoStatus(repo string) (status, details string) {
 		push.Output
 		Error string
 	}
-	if !(loadJSON(outputPath(repo, "push"), &pushOutput) == nil && pushOutput.Success) {
+	if !(loadJSON(outputPath(repoName, "push"), &pushOutput) == nil && pushOutput.Success) {
 		if pushOutput.Error != "" {
 			details = color.RedString("(push error) ") + pushOutput.Error
 		}
@@ -142,7 +136,8 @@ func getRepoStatus(repo string) (status, details string) {
 		merge.Output
 		Error string
 	}
-	if !(loadJSON(outputPath(repo, "merge"), &mergeOutput) == nil && mergeOutput.Success) {
+	// check PR was merged
+	if !(loadJSON(outputPath(repoName, "merge"), &mergeOutput) == nil && mergeOutput.Success) {
 		if mergeOutput.Error != "" {
 			details = color.RedString("(merge error) ") + mergeOutput.Error
 		}